CKW
1. 게임시작 시퀀스 추가
2. 각종 UI 수정
