+++ Assets/MEMBER/CKW/Scenes/CKW_GameStartScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4bb98c111aff42b4ba3ba556a04d3774
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
