+++ Assets/MEMBER/CKW/Scripts/CharacterSelectUI/UI_CharacterServerButton.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 08f4c0fddf9ea2e449233dbacc04ffbb
