+++ Assets/MEMBER/CKW/Scripts/TitleUI/GameStartupSequence.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ae299c12ac887b24dba25d0d4674d6e7
