--- Assets/MEMBER/CKW/Scripts/TitleUI/UI_CharacterServerButton.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 08f4c0fddf9ea2e449233dbacc04ffbb
